AL-10 Sort out synchronizing the rolling anon log and the harvester frequency

+++ collector/src/main/java/org/opendap/harvester/entity/document/HyraxInstance.java
@@ -20,7 +20,7 @@ public class HyraxInstance extends BaseEntity {
     @Getter @Setter
     private String reporterUrl;
     @Getter @Setter
-    private Integer ping = 0;
+    private Long ping = 0L;
     @Getter @Setter
     private Integer log = 0;
     @Getter @Setter
