+++ collector/src/main/java/org/opendap/harvester/entity/dto/model/RegisterModel.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.validator.constraints.URL;
 import org.springframework.beans.factory.annotation.Required;
+import org.springframework.web.bind.annotation.RequestParam;
 
 public class RegisterModel {
     @Getter @Setter
@@ -14,7 +15,7 @@ public class RegisterModel {
     @Getter @Setter
     private String reporterUrl;
     @Getter @Setter
-    private int ping;
+    private Long ping;
     @Getter @Setter
     private int log;
 }
